refactor(keybindings): extract src/keybindings/ into par-term-keybindings subcrate
Create new par-term-keybindings crate containing:
- lib.rs: KeybindingRegistry with from_config/lookup logic
- matcher.rs: KeybindingMatcher for winit event matching
- parser.rs: KeyCombo parser for "Ctrl+Shift+B" style strings

Dependencies: par-term-config (KeyBinding, ModifierRemapping), winit, log.

src/keybindings/mod.rs replaced with thin `pub use par_term_keybindings::*;`
re-export shim. matcher.rs and parser.rs deleted from main crate.

Fixed crate::config:: references in matcher.rs to use par_term_config::
directly.

Closes #177
