+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["par-term-acp", "par-term-config", "par-term-fonts", "par-term-render", "par-term-settings-ui", "par-term-ssh", "par-term-terminal"]
+members = ["par-term-acp", "par-term-config", "par-term-fonts", "par-term-keybindings", "par-term-render", "par-term-settings-ui", "par-term-ssh", "par-term-terminal"]
 resolver = "2"
 
 [package]
@@ -43,6 +43,9 @@ par-term-render = { path = "par-term-render", version = "0.1.0" }
 
 # Settings UI (egui-based settings interface)
 par-term-settings-ui = { path = "par-term-settings-ui", version = "0.1.0" }
+
+# Keybinding system (key parsing, matching, registry)
+par-term-keybindings = { path = "par-term-keybindings", version = "0.1.0" }
 
 # SSH host management and discovery
 par-term-ssh = { path = "par-term-ssh", version = "0.1.0" }
