+++ par-term-keybindings/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "par-term-keybindings"
+version = "0.1.0"
+edition = "2024"
+authors = ["Paul Robello <probello@gmail.com>"]
+description = "Keybinding system for par-term terminal emulator"
+license = "MIT"
+repository = "https://github.com/paulrobello/par-term"
+homepage = "https://github.com/paulrobello/par-term"
+keywords = ["terminal", "keybindings", "keyboard"]
+categories = ["gui"]
+
+[dependencies]
+# Configuration types (KeyBinding, ModifierRemapping)
+par-term-config = { path = "../par-term-config", version = "0.1.0" }
+
+# Window and input event types
+winit = "0.30"
+
+# Logging
+log = "0.4.29"
