+++ par-term-keybindings/src/parser.rs
